pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hydra_rhl"
version = "0.2"
description = "RHL kafka management tool"
readme = "README.md"
authors = [{ name = "Redhuntlabs", email = "rishabh.bhardwaj@redhuntlabs.com" }]
classifiers = [
    "Environment :: Web Environment",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8"
]
dependencies = ["dependency-injector >= 4.40.0", "kafka-python >= 2.0.2", 'python-dotenv >= 0.21.0']
requires-python = ">=3.6"

    [project.optional-dependencies]
    build = ["build", "twine"]
